requirements.txt
streamlit>=1.29.0
agno>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0